Check location in partition renaming
